Rename BaseServer classes
